Ajout du partial _form pour enlever la duplication de code de formulaire dans les templates

+++ src/Controller/PinsController.php
@@ -83,10 +83,13 @@ class PinsController extends AbstractController
     /**
      * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods="DELETE")
      */
-    public function delete(Pin $pin, EntityManagerInterface $em): Response
+    public function delete( Request $request, Pin $pin, EntityManagerInterface $em): Response
     {
-        $em->remove($pin);
-        $em->flush();
-        return $this->redirectToRoute('app_home');
+        $token = $request->request->get('csrf_token');
+        if ($this->isCsrfTokenValid('pin_deletion_'. $pin->getId(), $token)) {
+            $em->remove($pin);
+            $em->flush();
+            return $this->redirectToRoute('app_home');
+        }
     }
 }
